Refactoring climate devices
